requirements.txt
six==1.10.0
Flask==0.10.1
Flask-Cors==1.9.0
-e git+https://github.com/Beit-Hatfutsot/flask-security.git@develop#egg=Flask-Security
Flask-WTF==0.10.2
pyyaml==3.11
argparse==1.2.1
boto==2.35.1
flask-mongoengine==0.7.1
mongoengine==0.10.0
mongomock==3.2.1
pymongo==3.0.3
pytest>=2.7
pytest-flask==0.4.0
requests==2.11.1
soundcloud==0.4.1
gmail>=0.5
elasticsearch==1.4.0
redis==2.10.5
celery==3.1.23
pytest-mock==1.2
gcs-oauth2-boto-plugin==1.14
awesome-slugify==1.6.5
Pillow==3.3.1
